requirements.txt
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.1
fonttools==4.53.1
fsspec==2024.2.0
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.9.1
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.3
packaging==24.1
pillow==10.4.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
setuptools==70.0.0
six==1.16.0
sympy==1.12
tqdm==4.66.4
typing_extensions==4.9.0
